Update undo/redo to display the action that was undone/redone.
undo/redo previously displays "Action has been undone" or "Action
has been redone".

Undoable actions now commit a message to show the action that was
changed. The message is stored as a string in an array list
commandHistory, tied to the watcherStateList with the
currStatePointer. When undo/redo is called, the message is displayed
in the command result.